This makes acquire mostly/partly work in Simulate mode.
Exciting! But there’s still clearly some weird startup problem, in
which often the entire ring buffer gets read out the first time
through, yielding buffers of length [bufsize, 0], where the 2nd 0-size
buffer is definitely weird (or a consequence of a corner case when the
read and/or write pointers are at 0).