Failure to ConfigureLanceroSource does not cause a panic
If `ConfigureLanceroSource` fails (e.g., because there's no cringe configuration file), then remember this error but don't panic. For a non-Lancero computer, this failure is normal behavior.

Instead of panicking, remember the error and refuse to start the source in the `Sample()` method. (Given that a non-Lancero machine won't have `/dev/lancero_user*` files, this will not actually come up in practice.)

Fixes #140.