handle data drops for lancero source
When the first frame bit is in the incorrect location, we note a data drop. Then we estimate how many frames were dropped based on cpu time and sample rate. Then we write to a file ending in "data_drop.txt" with the framenum and estimated number of frames dropped. We also send a DATADROP message and flush the data drop file every 10 seconds. I modeled the code on how external triggers are handled to the extenet possible.