Make insertUniqueGetID stop incrementing IDs
It was silently auto-incrementing table id numbers when a given value was already in the table. By SELECTing first for the value and INSERTing only if it wasn't found, we avoid that problem.

Old way was always 2 DB transactions. New way is normally 1 (or 2 only in the case of adding a new value to the table).